+++ src/Stratis.Features.Collateral/MempoolRules/VotingRequestValidationRule.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.MemoryPool.Interfaces;
@@ -15,16 +13,19 @@ namespace Stratis.Bitcoin.Features.Collateral.MempoolRules
     {
         private readonly JoinFederationRequestEncoder encoder;
         private readonly VotingManager votingManager;
+        private readonly IFederationManager federationManager;
 
         public VotingRequestValidationRule(Network network,
             ITxMempool mempool,
             MempoolSettings mempoolSettings,
             ChainIndexer chainIndexer,
             ILoggerFactory loggerFactory,
-            VotingManager votingManager) : base(network, mempool, mempoolSettings, chainIndexer, loggerFactory)
+            VotingManager votingManager,
+            IFederationManager federationManager) : base(network, mempool, mempoolSettings, chainIndexer, loggerFactory)
         {
             this.encoder = new JoinFederationRequestEncoder(loggerFactory);
             this.votingManager = votingManager;
+            this.federationManager = federationManager;
         }
 
         /// <inheritdoc/>
@@ -51,58 +52,18 @@ namespace Stratis.Bitcoin.Features.Collateral.MempoolRules
                 PoAConsensusErrors.InvalidCollateralRequirement.Throw();
             }
 
+            if (!(this.federationManager is CollateralFederationManager federationManager))
+                return;
+
             // Prohibit re-use of collateral addresses.
             Script script = PayToPubkeyHashTemplate.Instance.GenerateScriptPubKey(request.CollateralMainchainAddress);
             string collateralAddress = script.GetDestinationAddress(this.network).ToString();
-            CollateralFederationMember owner = CollateralAddressOwner(VoteKey.AddFederationMember, collateralAddress);
+            CollateralFederationMember owner = federationManager.CollateralAddressOwner(this.votingManager, VoteKey.AddFederationMember, collateralAddress);
             if (owner != null && owner.PubKey != request.PubKey)
             {
                 this.logger.LogTrace("(-)[INVALID_COLLATERAL_REUSE]");
                 PoAConsensusErrors.VotingRequestInvalidCollateralReuse.Throw();
             }
-        }
-
-        private CollateralFederationMember GetMember(VotingData votingData)
-        {
-            if (!(this.network.Consensus.ConsensusFactory is CollateralPoAConsensusFactory collateralPoAConsensusFactory))
-                return null;
-
-            if (!(collateralPoAConsensusFactory.DeserializeFederationMember(votingData.Data) is CollateralFederationMember collateralFederationMember))
-                return null;
-
-            return collateralFederationMember;
-        }
-
-        public CollateralFederationMember CollateralAddressOwner(VoteKey voteKey, string address)
-        {
-            List<Poll> finishedPolls = this.votingManager.GetFinishedPolls();
-
-            CollateralFederationMember member = finishedPolls
-                .Where(x => !x.IsExecuted && x.VotingData.Key == voteKey)
-                .Select(x => this.GetMember(x.VotingData))
-                .FirstOrDefault(x => x.CollateralMainchainAddress == address);
-            
-            if (member != null)
-                return member;
-
-            List<Poll> pendingPolls = this.votingManager.GetPendingPolls();
-
-            member = pendingPolls
-                .Where(x => x.VotingData.Key == voteKey)
-                .Select(x => this.GetMember(x.VotingData))
-                .FirstOrDefault(x => x.CollateralMainchainAddress == address);
-
-            if (member != null)
-                return member;
-
-            List<VotingData> scheduledVotes = this.votingManager.GetScheduledVotes();
-
-            member = scheduledVotes
-                .Where(x => x.Key == voteKey)
-                .Select(x => this.GetMember(x))
-                .FirstOrDefault(x => x.CollateralMainchainAddress == address);
-
-            return member;
         }
     }
 }
